+++ tests/test_enterprise/test_signals.py
@@ -622,11 +622,6 @@ class TestEnterpriseLearnerRoleSignals(unittest.TestCase):
             enterprise_customer=self.second_enterprise_customer,
         )
 
-        enterprise_customer_user = EnterpriseCustomerUser.objects.filter(
-            user_id=self.learner_user.id
-        )
-        self.assertTrue(enterprise_customer_user.exists())
-
         # Verify that a new learner role assignment is created for both linked enterprise customers.
         learner_role_assignment = SystemWideEnterpriseUserRoleAssignment.objects.filter(
             user=self.learner_user,
@@ -642,15 +637,13 @@ class TestEnterpriseLearnerRoleSignals(unittest.TestCase):
         self.assertTrue(second_learner_role_assignment.exists())
 
         # Delete EnterpriseCustomerUser record.
+        enterprise_customer_user = EnterpriseCustomerUser.objects.get(
+            user_id=self.learner_user.id,
+            enterprise_customer=self.enterprise_customer,
+        )
         enterprise_customer_user.delete()
 
-        # Verify that enterprise_customer_user is deleted
-        enterprise_customer_user = EnterpriseCustomerUser.objects.filter(
-            user_id=self.learner_user.id
-        )
-        self.assertFalse(enterprise_customer_user.exists())
-
-        # Also verify that appropriate learner role assignment is deleted, but the other
+        # Verify appropriate learner role assignment is deleted, but the other
         # learner role assignment remains.
         learner_role_assignment = SystemWideEnterpriseUserRoleAssignment.objects.filter(
             user=self.learner_user,
@@ -658,12 +651,68 @@ class TestEnterpriseLearnerRoleSignals(unittest.TestCase):
             enterprise_customer=self.enterprise_customer,
         )
         self.assertFalse(learner_role_assignment.exists())
-        learner_role_assignment = SystemWideEnterpriseUserRoleAssignment.objects.filter(
+        second_learner_role_assignment = SystemWideEnterpriseUserRoleAssignment.objects.filter(
             user=self.learner_user,
             role=self.enterprise_learner_role,
             enterprise_customer=self.second_enterprise_customer,
         )
+        self.assertTrue(second_learner_role_assignment.exists())
+
+    def test_unlink_single_enterprise_learner_role_assignment(self):
+        """
+        Test that when `EnterpriseCustomerUser` record is unlinked, `assign_or_delete_enterprise_learner_role`
+        also deletes the enterprise learner role assignment associated with the user's previously linked enterprise
+        customer. However. it should *not* delete the enterprise learner role assignment associated with other
+        enterprise customers.
+        """
+        # Create a new EnterpriseCustomerUser record.
+        EnterpriseCustomerUserFactory(
+            user_id=self.learner_user.id,
+            enterprise_customer=self.enterprise_customer,
+        )
+
+        # Create a EnterpriseCustomerUser record such that the user is linked to multiple enterprises.
+        EnterpriseCustomerUserFactory(
+            user_id=self.learner_user.id,
+            enterprise_customer=self.second_enterprise_customer,
+        )
+
+        # Verify that a new learner role assignment is created for both linked enterprise customers.
+        learner_role_assignment = SystemWideEnterpriseUserRoleAssignment.objects.filter(
+            user=self.learner_user,
+            role=self.enterprise_learner_role,
+            enterprise_customer=self.enterprise_customer,
+        )
         self.assertTrue(learner_role_assignment.exists())
+        second_learner_role_assignment = SystemWideEnterpriseUserRoleAssignment.objects.filter(
+            user=self.learner_user,
+            role=self.enterprise_learner_role,
+            enterprise_customer=self.second_enterprise_customer,
+        )
+        self.assertTrue(second_learner_role_assignment.exists())
+
+        # Unlink EnterpriseCustomerUser record.
+        enterprise_customer_user = EnterpriseCustomerUser.objects.get(
+            user_id=self.learner_user.id,
+            enterprise_customer=self.enterprise_customer,
+        )
+        enterprise_customer_user.linked = False
+        enterprise_customer_user.save()
+
+        # Verify appropriate learner role assignment is deleted, but the other
+        # learner role assignment remains.
+        learner_role_assignment = SystemWideEnterpriseUserRoleAssignment.objects.filter(
+            user=self.learner_user,
+            role=self.enterprise_learner_role,
+            enterprise_customer=self.enterprise_customer,
+        )
+        self.assertFalse(learner_role_assignment.exists())
+        second_learner_role_assignment = SystemWideEnterpriseUserRoleAssignment.objects.filter(
+            user=self.learner_user,
+            role=self.enterprise_learner_role,
+            enterprise_customer=self.second_enterprise_customer,
+        )
+        self.assertTrue(second_learner_role_assignment.exists())
 
     def test_delete_enterprise_learner_role_assignment_no_role_assignment(self):
         """